Gemini/requirements.txt
Flask
python-dotenv
flask-cors
google-generativeai
sqlalchemy
pandas
psycopg2-binary